+++ sky/examples/stocks/lib/stock_menu.dart
@@ -4,7 +4,7 @@
 
 import 'package:sky/framework/components/popup_menu.dart';
 import 'package:sky/framework/fn.dart';
-import 'package:sky/framework/theme/view-configuration.dart';
+import 'package:sky/framework/theme/view_configuration.dart';
 
 class StockMenu extends Component {
   static final Style _style = new Style('''
